mgr/cephadm: add utils.py
move `name_to_config_section` and `assert_valid_host` into utils.py
