+++ src/pybind/mgr/cephadm/__init__.py
@@ -3,4 +3,4 @@ import os
 if 'UNITTEST' in os.environ:
     import tests
 
-from .module import CephadmOrchestrator, name_to_config_section
+from .module import CephadmOrchestrator
